requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.13.0
email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.2
apscheduler==3.10.4  
passlib[bcrypt,argon2]==1.7.4  
argon2-cffi==21.3.0 
bcrypt==4.0.1 
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
hiredis==2.2.3
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
python-dateutil==2.8.2
pytz==2023.3
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
hiredis==2.2.3
PyJwt 